修复库页封面视口崩溃，并落地元数据锁/刷新与阅读会话统一 (#52)
* 修复库页封面视口测量将 Sliver 强转为 RenderBox 导致的 TypeError。

EOF


* 修复系列详情漫画网格把 LayoutBuilder 当作 Sliver 子节点的崩溃。


* 将 FluentSelectField 改为自绘 Overlay 下拉，便于复用并统一表单选择体验。


* 阅读设置复用 FluentSelectField，并强制深色沉浸样式。


* 修复系列详情网格因路由名编码损坏导致无法打开漫画详情。


* 将阅读器自动播放开/关改为按 Comic 会话态，换本默认关闭。


* 为 Series read 退出回系列详情添加淡入微上浮转场。


* 将删除漫画确认框对齐为 HentaiDialog 风格并抽成共用组件。


* 收敛阅读热路径上的防御性 open_reader 调用（#49）。

取页/prefetch 改为要求已打开会话（ReaderSessionNotOpen）；会话引用计数与 with_ephemeral_reader 供缩略图冷路径；持锁读页优化见 #51。


* 修复从主页继续阅读时页码被可见索引过早回写而下漂。

Scroll/Paged/双页在 resume 对齐完成前忽略不匹配的 onPageChanged/可见页同步，避免进度被写成更低页码。


* 统一阅读会话为仅 comicId，废弃系列阅读进度（ADR-0005）。


* 缩小系列连载状态 chip，对齐同排 R18 的视觉密度。


* 支持系列内手动编辑排序序号（Komga 式锁定），并完善步进器交互。


* 移除系列编辑对话框漫画总数字段的占位提示。


* 为漫画与系列元数据添加 Komga 式字段锁，避免手动编辑与扫描覆盖冲突。


* 修复字段锁改动后的分析器报错与 Rust clippy 警告。


* 将 MultiSelect 改为上 label 下输入，使字段锁与图标在标签行右侧垂直居中对齐。


* 移除元数据字段锁按钮的 Tooltip，并清理对应文案。


* 保存元数据时仅对值变化的字段加锁，避免整表提交误锁未改字段。


* 统一漫画与系列编辑的成功/失败 Toast，不再区分字段级文案。


* 实现漫画/系列元数据刷新（Komga 式），详情页三点菜单入口，并与库同步互斥。


* 修复长按上下文菜单锚点：左上角对齐手势落点，并移除「跳转到详情页」。


* 修复 Android 因缺少 libc++_shared 导致 libhentai_flutter.so 无法加载。

unrar-ng 在 Android cdylib 上需显式链接并打包 C++ 运行时，否则 dlopen 会因 _ZTISt12length_error 等符号失败。


* 启用 Android PDF（pdfium）：打包 libpdfium、对齐桌面行为，并补上端到端 smoke 测试。

去掉 Android stub，vendor/CI/Gradle 接入 pdfium；升到 chromium/7825 以匹配 pdfium-render 所需符号。


* 补齐 PDF 行为回归测试，并让缺 vendor 时 build.rs 能可靠失效。

覆盖 parse 页数/元数据/零页跳过、阅读越界；HENTAI_VENDOR_DIR 变更会触发重新校验 pdfium。


* 集中 Metadata field lock：合并、Series 名/排序与自动上锁共用一处策略。

EOF


* 将 Resource 解析从 sync/parser 迁至 resource/parse。

EOF


* 拆出 resource module：切断 formats/reader/thumbnail 对 sync 解析的依赖。

EOF


* 加深 Library maintenance：互斥与会话清理留在 Rust，去掉 Dart 重复编排。

EOF


* 格式化 Dart 源码与测试，统一 dart format 风格。


---------

+++ app/test/domain/models/entity/comic/series_volume_labels_test.dart
@@ -13,7 +13,8 @@ Series _series({int itemCount = 3, int? totalCount}) {
     totalCount: totalCount,
     items: List<SeriesItem>.generate(
       itemCount,
-      (int index) => SeriesItem(comicId: 'comic-$index', order: index),
+      (int index) =>
+          SeriesItem(comicId: 'comic-$index', order: index.toDouble()),
     ),
   );
 }
